Bringing PEP8 back to above 8.
