+++ firestorm/urls.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from django.conf.urls import include, url
+from django.conf.urls import url
 
 from .views_user import home, TopicCreate
 from .views_kiosk import kiosk
